Fix the Get Asset by (Guid) Id returning Business Layer rather than Presentation layer entity.
This also fixes the bug, where the 'IsActive' Tenure field does not propagate to a response due to [IgnoreJson] attribute within Domain object. The lack of this field was causing a bug within the frontend, where the Tenure on the property page was being displayed as 'Innactive' while it was actually 'Active'.